requirements.txt
boto3
langchain
langchain-community
langchain-aws
pillow
streamlit
python-dotenv
pypdfium2
opencv-python
fpdf
openpyxl